requirements.txt
streamlit==1.39.0
pdfplumber==0.11.4
regex==2024.9.11
python-dotenv==1.0.1
plotly==5.24.1
unidecode==1.3.8
